+++ packages/cxalio-studio-tools/jpegger/filters/__init__.py
@@ -1,3 +1,8 @@
+"""Jpegger 图像过滤器集合。
+
+该包通过 star-import 汇聚所有公开过滤器，方便应用层统一引用。
+"""
+
 from .auto_resize import *
 from .colorspace import *
 from .image_filter import *